+++ planetmint/backend/models/dbtransaction.py
@@ -8,7 +8,6 @@ from dataclasses import dataclass, field
 
 from planetmint.backend.models import Asset, MetaData, Input, Script, Output
 
-
 @dataclass
 class DbTransaction:
     id: str = ""
@@ -43,15 +42,31 @@ class DbTransaction:
             inputs=Input.from_list_dict(transaction[5]),
             script=Script.from_dict(transaction[6]),
         )
+        
+    @staticmethod
+    def remove_generated_fields(tx_dict: dict):
+        tx_dict["outputs"] = [DbTransaction.remove_generated_or_none_output_keys(output) for output in tx_dict["outputs"]]
+        if "script" in tx_dict and tx_dict["script"] is None:
+            tx_dict.pop("script")
+        return tx_dict
+
+    @staticmethod
+    def remove_generated_or_none_output_keys(output):
+        output["condition"]["details"] = {k: v for k, v in output["condition"]["details"].items() if v is not None}
+        output.pop("id")
+        return output
+
 
     def to_dict(self) -> dict:
-        return {
-            "id": self.id,
-            "operation": self.operation,
-            "version": self.version,
+        tx = {
             "inputs": Input.list_to_dict(self.inputs),
             "outputs": Output.list_to_dict(self.outputs),
-            "assets": Asset.list_to_dict(self.assets),
+            "operation": self.operation,
             "metadata": self.metadata.to_dict() if self.metadata is not None else None,
-            "script": self.script.to_dict() if self.script is not None else None,
+            "assets": Asset.list_to_dict(self.assets),
+            "version": self.version,
+            "id": self.id,
+            "script": self.script.to_dict() if self.script is not None else None, 
         }
+        tx = DbTransaction.remove_generated_fields(tx)
+        return tx
